+++ app/src/main/java/com/cajetan/youtubeplus/viewmodels/SearchViewModel.kt
@@ -1,0 +1,23 @@
+package com.cajetan.youtubeplus.viewmodels
+
+import androidx.lifecycle.ViewModel
+import com.cajetan.youtubeplus.utils.FeedItem
+
+class SearchViewModel : ViewModel() {
+    var searchQuery: String = ""
+    var nextPageToken: String = ""
+
+    private var adapterItems: ArrayList<FeedItem> = ArrayList(emptyList())
+
+    fun getAdapterItems(): ArrayList<FeedItem> {
+        return adapterItems
+    }
+
+    fun addAdapterItems(items: List<FeedItem>) {
+        adapterItems.addAll(items)
+    }
+
+    fun clearAdapterItems() {
+        adapterItems = ArrayList(emptyList())
+    }
+}